refactor(db): drop shardChangeLogger from ShardLike's method set
